+++ aerith-common/Models/Competition.cs
@@ -9,10 +9,10 @@ namespace Aerith.Common.Models
     {
         // Properties
         [Column("groupId")]
-        public int GroupId { get; set; }
+        public long GroupId { get; set; }
 
         [Column("tournamentId")]
-        public int TournamentId { get; set; }
+        public long TournamentId { get; set; }
 
         // Navigation Properties
         [Required]
